+++ src/python-common/ceph/deployment/utils.py
@@ -6,25 +6,6 @@ from ceph.deployment.hostspec import SpecValidationError
 from numbers import Number
 
 
-def parse_combined_pem_file(pem_data: str) -> Tuple[Optional[str], Optional[str]]:
-
-    # Extract the certificate
-    cert_start = "-----BEGIN CERTIFICATE-----"
-    cert_end = "-----END CERTIFICATE-----"
-    cert = None
-    if cert_start in pem_data and cert_end in pem_data:
-        cert = pem_data[pem_data.index(cert_start):pem_data.index(cert_end) + len(cert_end)]
-
-    # Extract the private key
-    key_start = "-----BEGIN PRIVATE KEY-----"
-    key_end = "-----END PRIVATE KEY-----"
-    private_key = None
-    if key_start in pem_data and key_end in pem_data:
-        private_key = pem_data[pem_data.index(key_start):pem_data.index(key_end) + len(key_end)]
-
-    return cert, private_key
-
-
 def unwrap_ipv6(address):
     # type: (str) -> str
     if address.startswith('[') and address.endswith(']'):
